UNSTABLE - Refactoring to use helper methods
Also moving to a new "identifier" data system. This will be used for
generating statistic name tags based around the minecraft string id
system of mod:uniqueid convention. The helper methods are contained in
StatUtils and handle the grunt work of asyncronously updating the
statistic, reducing alot of the boilerplate code.